refactor: Remove obsolete data services and extract types
- Remove BaseDataService.ts (unused base class for legacy JSON services)
- Remove HeroDataService.ts (completely unused legacy JSON service)
- Extract DataService and IChangeTracked types to services/types.ts
- Update DatabaseHeroService to import types from new location
- Keep DatabaseHeroService.ts as it's actively used in 6 route files

All 405 tests passing after cleanup. Repository pattern is now the
primary data access layer with DatabaseHeroService as the service wrapper.
